Time: 15 ms (87.15%), Space: 17.3 MB (82.41%) - LeetHub

+++ 0018-4sum/0018-4sum.cpp
@@ -11,7 +11,7 @@ public:
                 int k=j+1;
                 int l=n-1;
                 while(k<l){
-                    long long sum=nums[i]+nums[j]+nums[k]+nums[l];
+                    long long sum=(long long)nums[i]+nums[j]+nums[k]+nums[l];
                     if(sum==target){
                         ans.push_back({nums[i],nums[j],nums[k],nums[l]});
                         k++;
